+++ 2021_RTCSA_dynfed/src/dynfed.rs
@@ -12,7 +12,7 @@ use lib::core::ProcessResult;
 use lib::graph_extension::{GraphExtension, NodeData};
 use lib::homogeneous::HomogeneousProcessor;
 use lib::processor::ProcessorBase;
-use lib::scheduler::*;
+use lib::scheduler_base::*;
 use lib::util::get_hyper_period;
 use lib::{getset_dag_set_scheduler, getset_dag_state_manager, log::*};
 use petgraph::{graph::NodeIndex, Graph};
