+++ 2021_RTCSA_dynfed/src/main.rs
@@ -8,7 +8,7 @@ use lib::graph_extension::GraphExtension;
 use lib::homogeneous::HomogeneousProcessor;
 use lib::log::dump_dag_set_scheduler_result_to_yaml;
 use lib::processor::ProcessorBase;
-use lib::scheduler::DAGSetSchedulerBase;
+use lib::scheduler_base::DAGSetSchedulerBase;
 use lib::util::{adjust_to_implicit_deadline, load_yaml};
 
 #[derive(Parser)]
