+++ lib/src/global_edf_scheduler.rs
@@ -4,10 +4,10 @@ use std::cmp::Ordering;
 
 use crate::getset_dag_set_scheduler;
 use crate::graph_extension::GraphExtension;
-use crate::scheduler::NodeDataWrapper;
+use crate::scheduler_base::NodeDataWrapper;
 use crate::{
     graph_extension::NodeData, homogeneous::HomogeneousProcessor, log::DAGSetSchedulerLog,
-    processor::ProcessorBase, scheduler::DAGSetSchedulerBase,
+    processor::ProcessorBase, scheduler_base::DAGSetSchedulerBase,
 };
 
 impl PartialOrd for NodeDataWrapper {
